+++ src/algotypes/sampling_algorithm.jl
@@ -85,3 +85,15 @@ Abstract super type for sample generators.
 """
 abstract type AbstractSampleGenerator end
 export AbstractSampleGenerator
+
+
+function bat_report!(md::Markdown.MD, generator::AbstractSampleGenerator)
+    alg = getalgorithm(generator)
+    if !(isnothing(alg) || ismissing(alg))
+        markdown_append!(md, """
+        ### Sample generation:
+
+        * Algorithm: $(nameof(typeof(alg)))
+        """)
+    end
+end