+++ NEWS.md
@@ -66,6 +66,8 @@ Breaking changes
 
 * New behavior of `ValueShapes.NamedTupleShape` and  `ValueShapes.NamedTupleDist`: Due to changes in [ValueShapes](https://github.com/oschulz/ValueShapes.jl) v0.10, `NamedTupleShape` and `NamedTupleDist` now either (by default) use `NamedTuple` or (optionally) `ValueShapes.ShapedAsNT`, but no longer a mix of them. As a result, the behavior of BAT has changed as well when using a `NamedTupleDist` as a prior. For example, `mode(samples).result` returns a `NamedTuple` now directly.
 
+* Use the new function `bat_report` to generate a sampling output report instead of `show(BAT.SampledDensity(samples))`.
+
 
 New experimental features
 -------------------------